Add files via upload
This uses CSS animations to create a layered scaling effect.  There are four seperate images layered on top of one another.  The two front mountain images scale in but, at different rates. The text image springs up from the bottom as it becomes opaque but never fully clears the mountain in front of it which gives it some perspective.

See the project at:

http://danielhunt.site/layered-effect/